requirements.txt
arrow==0.10.0
beautifulsoup4==4.5.3
Django==1.10.5
django-ckeditor==5.2.1
django-debug-toolbar==1.6
django-disqus==0.5
django-emoticons==1.2
django-forms-bootstrap==3.0.1
django-paypal==0.3.3
djangorestframework==3.5.2
djangorestframework-jwt==1.8.0
drfdocs==0.0.11
django-tinymce==2.6.0
funcsigs==1.0.2
gunicorn==19.6.0
mock==2.0.0
pbr==1.10.0
Pillow==2.9.0
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.2
stripe==1.46.0
whitenoise==3.2.3
